+++ src/components/8/About/AboutUsers.stories.js
@@ -1,20 +1,20 @@
-import About from "./About";
+import AboutUsers from "./AboutUsers";
 
 export default {
-    title: 'Components/8/About',
-    component: About,
+    title: 'Components/8/AboutUsers',
+    component: AboutUsers,
     argTypes: {
-      backgroundColor: { control: 'color' },
-      color: {control: 'color'},
-      Heading: {control: 'text'},
-      Content: {control: 'text'},
+        backgroundColor: { control: 'color' },
+        color: { control: 'color' },
+        Heading: { control: 'text' },
+        Content: { control: 'text' },
     },
-  };
+};
 
-  export const Default = (args) => (<About {...args} />)
+export const Default = (args) => (<AboutUsers {...args} />)
 
-  Default.args = {
+Default.args = {
     Heading: '',
     Content: 'Lorem ipsum dolor sit amet,  elit consectetur adipiscing elit. In nec tristique pharetra mi eu pellente. Morbi nec metus vel sem tristique porttitor. porta mauris ac odio nec suscipit pretium. Suspendisse maximus nunc ipsum, at gravida nunc posuere in.  vel sem tristique porttitor. porta mauris ac odio nec suscipit pretium. Suspendisse maximus nunc ipsum, at gravida nunc posuere in.',
 
-  }
+}
